+++ Cargo.toml
@@ -1,15 +1,7 @@
-[package]
-name = "scoped-mutex"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-critical-section = { version = "1.1", optional = true }
-
-[features]
-default = [
-    "impl-critical-section",
+[workspace]
+resolver = "2"
+members = [
+    "source/scoped-mutex",
+    "source/scoped-mutex-traits",
+    "source/scoped-mutex-impls",
 ]
-impl-critical-section = ["dep:critical-section"]
-impl-unsafe-cortex-m-thread-mode = []
-std = []
