+++ source/scoped-mutex-traits/Cargo.toml
@@ -1,0 +1,4 @@
+[package]
+name = "scoped-mutex-traits"
+version = "0.1.0"
+edition = "2021"
