+++ source/scoped-mutex/src/lib.rs
@@ -1,0 +1,4 @@
+#![cfg_attr(not(feature = "std"), no_std)]
+
+pub use scoped_mutex_traits::*;
+pub use scoped_mutex_impls::*;
